Introducing execution_context for thread_data
This patch replaces the use of the coroutine_self TLS pointer
by using the newly introduced execution_context for yield, suspend
and resume operations. What's missing is yield_to.

This has the side effect of not needing to check whether we are
running in an HPX thread when using synchronization primitives.